requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
feedparser==6.0.11
httpx==0.27.2
lxml==5.3.0
beautifulsoup4==4.12.3
pydantic==2.9.2
python-slugify==8.0.4
orjson==3.10.7
sqlite-utils==3.37
regex==2024.9.11
Unidecode==1.3.8
